// src/types/tmdbTypes.ts
export type Genre = {
  id: number;
  name: string;
};

export type Language = {
  iso_639_1: string;
  english_name: string;
  name: string;
};

export type Movie = {
  adult: boolean;
  backdrop_path: string | null;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string | null;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};